+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.10](https://github.com/realityforge/bazel-depgen/tree/v0.10) (2019-10-11) · [Full Changelog](https://github.com/realityforge/bazel-depgen/compare/v0.09...v0.10)
 
 * Upgrade the `org.realityforge.javax.annotation` artifact to version `1.0.1`.