+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.gir` artifact to version `0.11`.
+
 ### [v0.10](https://github.com/realityforge/bazel-depgen/tree/v0.10) (2019-10-11) · [Full Changelog](https://github.com/realityforge/bazel-depgen/compare/v0.09...v0.10)
 
 * Upgrade the `org.realityforge.javax.annotation` artifact to version `1.0.1`.